pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "upload-document-sdk"
version = "0.1.0"
description = "Document extraction and PDF filling SDK — open source"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.21",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.35.0"]
server = ["fastapi>=0.100.0", "uvicorn>=0.20.0"]
all = ["boto3>=1.35.0", "fastapi>=0.100.0", "uvicorn>=0.20.0"]
dev = ["pytest>=7.0", "pytest-cov", "black", "ruff"]

[project.urls]
Homepage = "https://github.com/yourorg/upload-document-sdk"
Repository = "https://github.com/yourorg/upload-document-sdk"

[tool.hatch.build.targets.wheel]
packages = ["src/uploaddocument"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: marks tests requiring API keys (deselect with '-m not integration')"]